Extract LobbyListAdapter
